Typescript implementation complete
