+++ src/components/trending/trending.data.tsx
@@ -1,0 +1,27 @@
+export const trendingItems = [
+	{
+		id: 56,
+		imageUrl: "https://i.ibb.co/Hr1CyZ5/Burbank-Shawl.webp",
+		name: "Burbank Shawl",
+		price: 1835,
+	},
+
+	{
+		id: 31,
+		imageUrl: "https://i.ibb.co/VTtQWtL/Static-Nike.webp",
+		name: "Static Nike",
+		price: 12500,
+	},
+	{
+		id: 35,
+		imageUrl: "https://i.ibb.co/1rrrxd7/Cotton-Suit-Blazer.webp",
+		name: "Cotton Suit Blazer",
+		price: 3770,
+	},
+	{
+		id: 7,
+		imageUrl: "https://i.ibb.co/r733R0W/Burbank-Black.webp",
+		name: "Burbank Black",
+		price: 499,
+	},
+];
